pyproject.toml
[tool.poetry]
name = "course-work-7"
version = "0.1.0"
description = ""
authors = ["Oleg Smirnov <oleg.smirnoff505@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
pillow = "^10.2.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.1"
django-cors-headers = "^4.3.1"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.4.3"
flake8 = "^7.0.0"
flake8-html = "^0.4.3"
drf-yasg = "^1.21.7"
celery = "^5.3.6"
django-celery-beat = "^2.6.0"
redis = "^5.0.2"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
